requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==3.0.2
Django==4.2
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-filter==23.1
django-phonenumber-field==7.0.2
django-redis==5.2.0
django-render-block==0.9.2
django-templated-email==3.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
msgpack==1.0.5
packaging==23.1
phonenumberslite==8.13.10
Pillow==9.5.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pytz==2023.3
redis==4.5.4
requests==2.28.2
ruamel.yaml==0.17.21
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
